requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pandas
scikit-learn
